perampingan pada component
