--- src/components/album/index.js
@@ -1,5 +1,0 @@
-const Album = ({ album }) => {
-	return <p>{album}</p>;
-};
-
-export default Album;
