--- src/components/artist/index.js
@@ -1,5 +1,0 @@
-const Artist = ({ artist }) => {
-	return <p>{artist}</p>;
-};
-
-export default Artist;
