+++ src/components/image/index.js
@@ -1,5 +1,5 @@
 const ImageMusic = ({ image }) => {
-	return <img src={image} alt="paman rock" />;
+	return <img src={image} alt="paman mario" />;
 };
 
 export default ImageMusic;
