+++ src/components/song/index.js
@@ -1,5 +1,5 @@
-const Song = ({ title }) => {
-	return <p>{title}</p>;
+const Song = ({ title, artist, album }) => {
+	return (<p>{title}</p>), (<p>{artist}</p>), (<p>{album}</p>);
 };
 
 export default Song;
